Fix Issue That The Thread For The Game Not Responding
In Host and Client, the startGame() method is called in a new thread (here we name it game thread). Wherein, a while loop is used to wait for the current user to select a button.

This causes a problem that if nothing being done in this while loop (e.g., continually printing something) this game thread doesn't continue executing, i.e., sending data to the opponent about which button it has selected, eventhough the user may have selected a button on the GUI.

	modified:   Host/my-app/src/main/java/com/curtisnewbie/app/Host.java
	modified:   Client/my-app/src/main/java/com/curtisnewbie/app/Client.java