Fix Issue Where Program Doesn't Check Winning Condition
Winning condition and the condition where all buttons are selected are supposed to be checked once the user or opponent moves. However, due to some problems in multithreading, it is not operating as expected.

This is fixed by having the Platform.runLater() method covering all the operation in opponentMoveTo() and moveTo() methods.

	modified:   src/main/java/com/curtisnewbie/app/GamePane.java