Update GamePane Related To The Use of frezze() and unfreeze()
Previously, methods in GamePane() calls the freeze() and unfreeze() methods to disable or make the GamePane itself available. This should be controlled by the controller (Host class and Clent class) instead of the GUI itself. So, the methods like moveTo() and opponentMoveTo() no longer call the freeze() and unfreeze() methods internally.

	modified:   src/main/java/com/curtisnewbie/app/GamePane.java