Client: Handle the situation where it cannot connect to host
Before this modification, the thread continues the start the game even through the IOException is being thrown when the connection fails. This inappropriate behaviour is corrected by this modification; after this modification, a dialog is shown when connection fails and the thread will not continue operating the game.

This modification is only for Client rather than for Host, since Host is always waiting for connection.

	modified:   Client/my-app/src/main/java/com/curtisnewbie/app/Client.java